# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ASR Service API
  description: API for Audio Speech Recognition and Task Management
  version: 1.0.0
servers:
  - url: http://localhost:7200
    description: Local development server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
...
    Permission:
      type: string
      enum: [Transcribe]
      description: Available permissions for API keys
    RateLimit:
      type: object
      required:
        - requests_per_minute
        - burst_size
      properties:
        requests_per_minute:
          type: integer
          minimum: 1
          description: Maximum number of requests allowed per minute
        burst_size:
          type: integer
          minimum: 1
          description: Maximum burst size for rate limiting
    CreateApiKeyRequest:
      type: object
      required:
        - name
        - permissions
        - rate_limit
      properties:
        name:
          type: string
          description: Name of the API key
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
          description: List of permissions for the API key
        rate_limit:
          $ref: '#/components/schemas/RateLimit'
        expires_in_days:
          type: integer
          nullable: true
          description: Number of days until the API key expires
    ApiKeyInfo:
      type: object
      required:
        - api_key
        - name
        - permissions
        - rate_limit
        - created_at
      properties:
        api_key:
          type: string
          description: The generated API key
        name:
          type: string
          description: Name of the API key
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        rate_limit:
          $ref: '#/components/schemas/RateLimit'
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
...
paths:
...
  /auth/api-keys:
    post:
      summary: Create a new API key
      description: Creates a new API key with specified permissions and rate limits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              name: "My API Key"
              permissions: ["Transcribe"]
              rate_limit:
                requests_per_minute: 60
                burst_size: 10
              expires_in_days: 30
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          key_info:
                            $ref: '#/components/schemas/ApiKeyInfo'
              example:
                success: true
                data:
                  key_info:
                    api_key: "sk-xxxxxxxxxxxxxxxxxxxx"
                    name: "My API Key"
                    permissions: ["Transcribe"]
                    rate_limit:
                      requests_per_minute: 60
                      burst_size: 10
                    created_at: "2024-03-15T10:00:00Z"
                    expires_at: "2024-04-14T10:00:00Z"
                error: null
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: false
                data: null
                error: "Invalid rate limit configuration"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: false
                data: null
                error: "Internal server error"
  /auth/api-keys/{api_key}:
    delete:
      summary: Revoke an API key
      description: Revokes (deletes) an existing API key
      parameters:
        - name: api_key
          in: path
          required: true
          schema:
            type: string
          description: The API key to revoke
      responses:
        '200':
          description: API key revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data: null
                error: null
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: false
                data: null
                error: "API key not found"
...